updated the changes requested by Riaan
fixed the following feedback from Riaan.

"
A question when i touch the screen I can move the website a bit left and a bit right is that fine like that.

And it only downloaded half of the pictures the others is just a block. Are you still working on it.

Otherwise it is awesome so far.

PS: the first three pics shows no pictures.

       the last two is when I move the screen to the left and to the right and some
       of the letters disappears.

"

"
Hi Y,

It is looking so cool.

Okay where it says - Custom Designs - Zippo type lighters and bags (Take the bags word out)

Where it says FISHING HOODIES - Spelling error - hooie (hoodie)

Take the sentence - The hood drawcord is standard with eyelets - take this sentence out please.

The size is fine for my laptop on the screen.

And that is that. The rest is awesome as it is.
"